Refactor ColorsLab
