+++ PowerPointLabs/PowerPointLabs/ColorsLab/Magnifier.xaml
@@ -1,9 +1,8 @@
-﻿<Window x:Class="PowerPointLabs.ColorPicker.Magnifier"
+﻿<Window x:Class="PowerPointLabs.ColorsLab.Magnifier"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-        xmlns:local="clr-namespace:PowerPointLabs.ColorPicker"
         Name="MagnifierMainWindow"
         AllowsTransparency="True"
         Loaded="Window_Loaded"
