+++ PowerPointLabs/PowerPointLabs/ColorsLab/MagnifierOverlay.xaml
@@ -1,9 +1,8 @@
-﻿<Window x:Class="PowerPointLabs.ColorPicker.MagnifierOverlay"
+﻿<Window x:Class="PowerPointLabs.ColorsLab.MagnifierOverlay"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:PowerPointLabs.ColorPicker"
              Name="MagnifierMainWindow"
              AllowsTransparency="True"
              ShowInTaskbar="False"
