+++ PowerPointLabs/PowerPointLabs/Converters/ColorPane/HSLColorToHex.cs
@@ -2,6 +2,8 @@
 using System.Drawing;
 using System.Windows.Data;
 
+using PowerPointLabs.ColorsLab;
+
 namespace PowerPointLabs.Converters.ColorPane
 {
     [ValueConversion(typeof(HSLColor), typeof(string))]
@@ -12,7 +14,7 @@ namespace PowerPointLabs.Converters.ColorPane
         public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {
             Color color = (HSLColor)value;
-            return "#" + color.R.ToString("X2") + color.G.ToString("X2") + color.B.ToString("X2");
+            return ColorHelper.ColorToHexString(color);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
