+++ PowerPointLabs/PowerPointLabs/Converters/ColorPane/SelectedColorShiftLuminosityByFactor.cs
@@ -2,6 +2,8 @@
 using System.Drawing;
 using System.Windows.Data;
 
+using PowerPointLabs.ColorsLab;
+
 namespace PowerPointLabs.Converters.ColorPane
 {
     [ValueConversion(typeof(HSLColor), typeof(string))]
@@ -14,12 +16,12 @@ namespace PowerPointLabs.Converters.ColorPane
             if (value == null)
             {
                 Color color = (HSLColor)value;
-                return "#" + color.R.ToString("X2") + color.G.ToString("X2") + color.B.ToString("X2");
+                return ColorHelper.ColorToHexString(color);
             }
             HSLColor selectedColor = (HSLColor)value;
             float shiftFactor = float.Parse((string)parameter);
             Color convertedColor = new HSLColor(selectedColor.Hue, selectedColor.Saturation, shiftFactor * 240);
-            return "#" + convertedColor.R.ToString("X2") + convertedColor.G.ToString("X2") + convertedColor.B.ToString("X2");
+            return ColorHelper.ColorToHexString(convertedColor);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
